:wave: Greet the visitor like it is 1999
Zoe Quinn's book Crash Override mentioned the lost 90s internet youthful exuberance of websites personally greeting the visitor. The current design of the page is a direct retro throwback to 1999, so I felt inclined to add a personal welcome message to fit. The irony is not lost on me that the 1999 archive did not have a Welcome message in its weird teen me way of trying to seem semi-professional in my own bizarre way.